Start passing tests again
