+++ tests/integration/cmip7-scenariomip/test_integration_cmip7_scenariomip_pre_processing.py
@@ -6,8 +6,6 @@ import itertools
 import re
 from functools import partial
 
-import numpy as np
-import pandas as pd
 import pytest
 from pandas_openscm.index_manipulation import update_index_levels_func
 
@@ -15,214 +13,24 @@ from gcages.cmip7_scenariomip import CMIP7ScenarioMIPPreProcessor
 
 # from gcages.cmip7_scenariomip.pre_processing import InternalConsistencyError
 from gcages.renaming import SupportedNamingConventions, convert_variable_name
-from gcages.testing import assert_frame_equal
+from gcages.testing import (
+    assert_frame_equal,
+    get_cmip7_scenariomip_like_input,
+    stack_sector_and_return_to_variable,
+    unstack_sector,
+)
 from gcages.units_helpers import strip_pint_incompatible_characters_from_units
 
-git = pytest.importorskip("git")
-nomenclature = pytest.importorskip("nomenclature")
 pix = pytest.importorskip("pandas_indexing")
-
-RNG = np.random.default_rng()
-
-
-def aggregate_up_sectors(indf, copy=False):
-    res = indf
-    if copy:
-        res = res.copy()
-
-    aggregations = set(["|".join(c.split("|")[:-1]) for c in res if "|" in c])
-    # Have to aggregate lowest level sectors first
-    aggregations_sorted = sorted(aggregations, key=lambda x: x.count("|"))[::-1]
-    for aggregation in aggregations_sorted:
-        if aggregation in res:
-            msg = f"{aggregation} already in indf?!"
-            raise KeyError(msg)
-
-        contributing = []
-        for c in res:
-            if not c.startswith(f"{aggregation}|"):
-                continue
-
-            split = c.split(f"{aggregation}|")
-            if "|" in split[-1]:
-                # going too many levels deep
-                continue
-
-            contributing.append(c)
-
-        res[aggregation] = res[contributing].sum(axis="columns")
-
-    return res
-
-
-def split_variable(df):
-    res = df.pix.extract(variable="{table}|{gas}|{sector}")
-
-    return res
-
-
-def combine_to_make_variable(df):
-    res = df.pix.format(variable="{table}|{gas}|{sector}", drop=True)
-
-    return res
-
-
-def add_gas_totals(indf):
-    # Should be called after aggregate_up_sectors
-
-    top_level = [c for c in indf if "|" not in c]
-
-    sector_stuff = combine_to_make_variable(
-        indf.unstack("year").stack("sector", future_stack=True)
-    )
-    gas_totals = (
-        indf[top_level]
-        .sum(axis="columns")
-        .unstack("year")
-        .pix.format(variable="{table}|{gas}", drop=True)
-    )
-
-    res = pix.concat([sector_stuff, gas_totals])
-
-    return res
 
 
 @pytest.fixture(scope="session")
 def example_raw_input():
-    bottom_level_sectors = [
-        # Energy sector
-        "Energy|Supply",
-        # Industrial sector stuff
-        "Energy|Demand|Industry",
-        "Energy|Demand|Other Sector",
-        "Industrial Processes",
-        "Other",
-        # Residential commercial and other
-        "Energy|Demand|Residential and Commercial and AFOFI",
-        # Solvents production and application
-        "Product Use",
-        # Aviation stuff
-        "Energy|Demand|Transportation|Domestic Aviation",
-        # something so we can get "Energy|Demand|Transportation" as a sum
-        "Energy|Demand|Transportation|Rail",
-        "Energy|Demand|Bunkers|International Aviation",
-        # International shipping
-        "Energy|Demand|Bunkers|International Shipping",
-        # Waste
-        "Waste",
-        # Agriculture
-        "AFOLU|Agriculture",
-        # Burning sectors
-        "AFOLU|Agricultural Waste Burning",
-        "AFOLU|Land|Fires|Forest Burning",
-        "AFOLU|Land|Fires|Grassland Burning",
-        "AFOLU|Land|Fires|Peat Burning",
-        # Imperfect but put these in to test agriculture aggregation too for now
-        # (we don't have a better source to harmonise too,
-        # except for CO2 but they don't use CO2 LULUCF
-        # emissions as input anyway, they use land-use change patterns).
-        # (For SCMs, this also doesn't matter as it all gets rolled up to
-        # "AFOLU" anyway because SCMs aren't able to handle the difference,
-        # except maybe OSCAR but that isn't in OpenSCM-Runner
-        # so we can't guess anything to do with OSCAR for now anyway.)
-        "AFOLU|Land|Land Use and Land-Use Change",
-        "AFOLU|Land|Harvested Wood Products",
-        "AFOLU|Land|Other",
-        "AFOLU|Land|Wetlands",
-    ]
-
-    timesteps = np.arange(2015, 2100 + 1, 5)
-    start_index = pd.MultiIndex.from_product(
-        [
-            ["model_a"],
-            ["scenario_a"],
-            ["Emissions"],
-            [
-                "CO2",
-                "CH4",
-                "N2O",
-                "BC",
-                "CO",
-                "NH3",
-                "OC",
-                "NOx",
-                "Sulfur",
-                "VOC",
-            ],
-            ["model_a|China", "model_a|Pacific OECD"],
-            timesteps,
-        ],
-        names=["model", "scenario", "table", "gas", "region", "year"],
-    )
-    df = pd.DataFrame(
-        RNG.random(start_index.shape[0]),
-        columns=pd.Index(["Other"], name="sector"),
-        index=start_index,
-    )
-
-    for bls in bottom_level_sectors:
-        df[bls] = RNG.random(df.index.shape[0])
-
-    df = aggregate_up_sectors(df)
-    df = add_gas_totals(df)
-
-    def get_unit(v):
-        species = v.split("|")[1]
-        unit_map = {
-            "BC": "Mt BC/yr",
-            "CH4": "Mt CH4/yr",
-            "CO": "Mt CO/yr",
-            "CO2": "Mt CO2/yr",
-            "N2O": "kt N2O/yr",
-            "NH3": "Mt NH3/yr",
-            "NOx": "Mt NO2/yr",
-            "OC": "Mt OC/yr",
-            "Sulfur": "Mt SO2/yr",
-            "VOC": "Mt VOC/yr",
-        }
-        return unit_map[species]
-
-    df["unit"] = df.index.get_level_values("variable").map(get_unit)
-    df = df.set_index("unit", append=True)
-    df = pix.concat(
-        [
-            df.groupby(df.index.names.difference(["region"]))
-            .sum()
-            .pix.assign(region="World"),
-            df,
-        ]
-    )
-    global_only_base = pd.DataFrame(
-        RNG.random(timesteps.size)[np.newaxis, :],
-        columns=df.columns,
-        index=start_index.droplevel(["region", "year", "gas"]).drop_duplicates(),
-    )
-    global_only_l = []
-    for global_only_gas, unit in [
-        ("HFC|HFC23", "kt HFC23/yr"),
-        ("HFC", "kt HFC134a-equiv/yr"),
-        ("HFC|HFC134a", "kt HFC134a/yr"),
-        ("HFC|HFC43-10", "kt HFC43-10/yr"),
-        ("PFC", "kt CF4-equiv/yr"),
-        ("F-Gases", "Mt CO2-equiv/yr"),
-        ("SF6", "kt SF6/yr"),
-        ("CF4", "kt CF4/yr"),
-        ("C2F6", "kt C2F6/yr"),
-        ("C6F14", "kt C6F14/yr"),
-    ]:
-        global_only_l.append(
-            global_only_base.pix.assign(gas=global_only_gas, unit=unit, region="World")
-            .pix.format(variable="{table}|{gas}", drop=True)
-            .reorder_levels(df.index.names)
-        )
-
-    df = pix.concat([df, *global_only_l])
-
-    return df
+    return get_cmip7_scenariomip_like_input()
 
 
 @pytest.fixture(scope="session")
-def processed_output(example_raw_input, default_data_structure_definition):
+def processed_output(example_raw_input):
     pre_processor = CMIP7ScenarioMIPPreProcessor(
         n_processes=None,  # run serially
     )
@@ -241,7 +49,7 @@ def test_transport_shuffling(example_raw_input, processed_output):
     subtracted to make |Transportation Sector.
     """
     # Not interested in global level for this
-    df_to_check = processed_output.region_sector_workflow_emissions
+    df_to_check = processed_output.gridding_workflow_emissions
 
     # The original Transportation should be dropped out
     assert (
@@ -250,41 +58,29 @@ def test_transport_shuffling(example_raw_input, processed_output):
         .any()
     ), df_to_check.pix.unique("variable")
 
-    example_raw_input_sectors = split_variable(example_raw_input)
+    example_raw_input_sectors = unstack_sector(example_raw_input)
 
-    exp_aircraft = combine_to_make_variable(
-        example_raw_input_sectors.loc[
-            pix.ismatch(
-                sector=[
-                    "Energy|Demand|Transportation|Domestic Aviation",
-                    "Energy|Demand|Bunkers|International Aviation",
-                ],
-                gas=["CO2", "CH4"],
-                region="model_a**",
-            )
+    tmp = example_raw_input_sectors.copy()
+    tmp["Aircraft"] = tmp[
+        [
+            "Energy|Demand|Transportation|Domestic Aviation",
+            "Energy|Demand|Bunkers|International Aviation",
         ]
-        .groupby(example_raw_input_sectors.index.names.difference(["sector"]))
-        .sum()
-        .pix.assign(sector="Aircraft")
-    ).reorder_levels(example_raw_input.index.names)
-
-    example_raw_input_sectors_stacked = (
-        example_raw_input_sectors.loc[
-            pix.ismatch(gas=["CO2", "CH4"], region="model_a**")
-        ]
-        .unstack("sector")
-        .stack("year", future_stack=True)
+    ].sum(axis="columns", min_count=2)
+    exp_aircraft = stack_sector_and_return_to_variable(
+        # Only expect results reported at the World level
+        tmp[["Aircraft"]].dropna().loc[pix.ismatch(region="World")]
     )
-    exp_transportation_sector = combine_to_make_variable(
-        (
-            example_raw_input_sectors_stacked["Energy|Demand|Transportation"]
-            - example_raw_input_sectors_stacked[
-                "Energy|Demand|Transportation|Domestic Aviation"
-            ]
-        )
-        .unstack("year")
-        .pix.assign(sector="Transportation Sector")
-    ).reorder_levels(example_raw_input.index.names)
+
+    tmp = example_raw_input_sectors.copy()
+    tmp["Transportation Sector"] = (
+        tmp["Energy|Demand|Transportation"]
+        - tmp["Energy|Demand|Transportation|Domestic Aviation"]
+    )
+    exp_transportation_sector = stack_sector_and_return_to_variable(
+        # Only expect results reported at the regional level
+        tmp[["Transportation Sector"]].dropna().loc[~pix.ismatch(region="World")]
+    )
 
     assert_frame_equal(
         df_to_check.loc[pix.ismatch(variable="**Aircraft")],
@@ -298,7 +94,8 @@ def test_transport_shuffling(example_raw_input, processed_output):
 
 
 def test_industrial_sector_aggregation(example_raw_input, processed_output):
-    df_to_check = processed_output.region_sector_workflow_emissions
+    # Not interested in global level for this
+    df_to_check = processed_output.gridding_workflow_emissions
 
     exp_sector = "Industrial Sector"
     exp_contributing_sectors = [
@@ -308,15 +105,15 @@ def test_industrial_sector_aggregation(example_raw_input, processed_output):
         "Other",
     ]
 
-    example_raw_input_sectors = split_variable(example_raw_input)
-    exp_sector_df = combine_to_make_variable(
-        example_raw_input_sectors.loc[
-            pix.ismatch(sector=exp_contributing_sectors, region="model_a**")
-        ]
-        .groupby(example_raw_input_sectors.index.names.difference(["sector"]))
-        .sum()
-        .pix.assign(sector=exp_sector)
-    ).reorder_levels(example_raw_input.index.names)
+    example_raw_input_sectors = unstack_sector(example_raw_input)
+    tmp = example_raw_input_sectors.copy()
+    tmp[exp_sector] = tmp[exp_contributing_sectors].sum(
+        axis="columns", min_count=len(exp_contributing_sectors)
+    )
+    exp_sector_df = stack_sector_and_return_to_variable(
+        # Only expect results reported at the regional level
+        tmp[[exp_sector]].dropna().loc[~pix.ismatch(region="World")]
+    )
 
     assert_frame_equal(
         df_to_check.loc[pix.ismatch(variable=f"**{exp_sector}")],
@@ -325,7 +122,8 @@ def test_industrial_sector_aggregation(example_raw_input, processed_output):
 
 
 def test_agricultural_sector_aggregation(example_raw_input, processed_output):
-    df_to_check = processed_output.region_sector_workflow_emissions
+    # Not interested in global level for this
+    df_to_check = processed_output.gridding_workflow_emissions
 
     exp_sector = "Agriculture"
     exp_contributing_sectors = [
@@ -336,15 +134,15 @@ def test_agricultural_sector_aggregation(example_raw_input, processed_output):
         "AFOLU|Land|Wetlands",
     ]
 
-    example_raw_input_sectors = split_variable(example_raw_input)
-    exp_sector_df = combine_to_make_variable(
-        example_raw_input_sectors.loc[
-            pix.ismatch(sector=exp_contributing_sectors, region="model_a**")
-        ]
-        .groupby(example_raw_input_sectors.index.names.difference(["sector"]))
-        .sum()
-        .pix.assign(sector=exp_sector)
-    ).reorder_levels(example_raw_input.index.names)
+    example_raw_input_sectors = unstack_sector(example_raw_input)
+    tmp = example_raw_input_sectors.copy()
+    tmp[exp_sector] = tmp[exp_contributing_sectors].sum(
+        axis="columns", min_count=len(exp_contributing_sectors)
+    )
+    exp_sector_df = stack_sector_and_return_to_variable(
+        # Only expect results reported at the regional level
+        tmp[[exp_sector]].dropna().loc[~pix.ismatch(region="World")]
+    )
 
     assert_frame_equal(
         df_to_check.loc[pix.ismatch(variable=f"**{exp_sector}")],
@@ -353,7 +151,7 @@ def test_agricultural_sector_aggregation(example_raw_input, processed_output):
 
 
 def test_output_sectors(example_raw_input, processed_output):
-    example_raw_input_sectors = split_variable(example_raw_input)
+    example_raw_input_sectors = unstack_sector(example_raw_input)
 
     exp_output_sectors = [
         # Fossil
@@ -378,13 +176,15 @@ def test_output_sectors(example_raw_input, processed_output):
         f"{table}|{gas}|{sector}"
         for table, gas, sector in itertools.product(
             example_raw_input_sectors.pix.unique("table"),
-            ["CO2", "CH4"],
+            example_raw_input_sectors.loc[~pix.isin(region="World")].pix.unique(
+                "species"
+            ),
             exp_output_sectors,
         )
     ]
 
     assert set(exp_output_variables) == set(
-        processed_output.region_sector_workflow_emissions.pix.unique("variable")
+        processed_output.gridding_workflow_emissions.pix.unique("variable")
     )
